+++ js/engine.js
@@ -217,7 +217,7 @@ function collisionChecks(event) {
                         if (obj.classType === "body" && obj.speed > 6) {
                             const v = Vector.magnitude(Vector.sub(mob[k].velocity, obj.velocity));
                             if (v > 9) {
-                                let dmg = tech.blockDamage * m.dmgScale * v * obj.mass * (tech.isMobBlockFling ? 2.5 : 1) * (tech.isBlockRestitution ? 2.5 : 1);
+                                let dmg = tech.blockDamage * m.dmgScale * v * obj.mass * (tech.isMobBlockFling ? 2.5 : 1) * (tech.isBlockRestitution ? 2.5 : 1) * (tech.isPWBlockDecay ? 3 : 1);
                                 if (mob[k].isShielded) dmg *= 0.7
                                 // console.log(dmg)
                                 mob[k].damage(dmg, true);